+++ sso/user/models.py
@@ -1,13 +1,11 @@
 from django.conf import settings
+from django.contrib.auth.models import AbstractBaseUser, BaseUserManager, PermissionsMixin
+from django.contrib.postgres.fields import JSONField
 from django.core.mail import send_mail
 from django.db import models
-from django.contrib.auth.models import (
-    AbstractBaseUser, BaseUserManager, PermissionsMixin
-)
-from django.contrib.postgres.fields import JSONField
+from django.urls import reverse
 from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
-from django.urls import reverse
 
 from allauth.account.forms import default_token_generator, user_pk_to_url_str
 from allauth.account.models import EmailConfirmation
@@ -159,29 +157,11 @@ class UserProfile(TimeStampedModel):
     class Meta:
         ordering = ['-created']
 
-    user = models.OneToOneField(
-        User,
-    )
-
-    first_name = models.CharField(
-        max_length=128,
-    )
-
-    last_name = models.CharField(
-        max_length=128,
-    )
-
-    job_title = models.CharField(
-        max_length=128,
-        blank=True,
-        null=True
-    )
-
-    mobile_phone_number = models.CharField(
-        max_length=128,
-        blank=True,
-        null=True,
-    )
+    user = models.OneToOneField(User, related_name='user_profile')
+    first_name = models.CharField(max_length=128)
+    last_name = models.CharField(max_length=128)
+    job_title = models.CharField(max_length=128, blank=True, null=True)
+    mobile_phone_number = models.CharField(max_length=128, blank=True, null=True)
 
     def __str__(self):
         return str(self.user)