+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - No ticket - Upgrade django to 1.11.23 for vulnerability fix
 - No ticket - Fix vulnerability with yarn
 - No ticket - Cleared up old settings
+- TT-1758 - Fix breadcrumbs on change password page
 
 ## [2019.08.13](https://github.com/uktrade/directory-sso/releases/tag/2019.08.13_1)
 [Full Changelog](https://github.com/uktrade/directory-sso/compare/2019.08.13...2019.08.13_1)