+++ CHANGELOG.md
@@ -3,6 +3,28 @@
 ## Pre release
 
 ### Implemented enhancements
+- No ticket - Added django settings janitor
+- Expose user profile on user retrieve endpoint
+- TT-1673 - Redirect to enrolment after login if no company or personal details
+- TT-1760 - added user profile update
+
+### Fixed bugs:
+- No ticket - Upgrade django to 1.11.23 for vulnerability fix
+- No ticket - Fix vulnerability with yarn
+- No ticket - Cleared up old settings
+- TT-1758 - Fix breadcrumbs on change password page
+- TT-1779 - Fix login hanging
+
+## [2019.08.13](https://github.com/uktrade/directory-sso/releases/tag/2019.08.13_1)
+[Full Changelog](https://github.com/uktrade/directory-sso/compare/2019.08.13...2019.08.13_1)
+
+### Hot fix
+- No ticket - Add feature flag to "start now" on sign in page
+
+## [2019.08.13](https://github.com/uktrade/directory-sso/releases/tag/2019.08.13)
+[Full Changelog](https://github.com/uktrade/directory-sso/compare/2019.07.09...2019.08.13)
+
+### Implemented enhancements
 - No Ticket - Adding missing models to admin view
 - TT-990 -  Making job title optional in user profile to accomodate individual user profile.
 - TT-1573 - Expose has_user_profile on session user retrieval